backend/requirements.txt
asyncpg==0.30.0
aiosqlite==0.21.0
alembic==1.14.1
bcrypt==3.2.2
fastapi==0.115.8
greenlet==3.1.1
passlib==1.7.4
pydantic_settings==2.8.1
pydantic==2.10.6
pytest==8.3.4
pytest-asyncio==0.25.3
python-jose[cryptography]
python-multipart
SQLAlchemy==2.0.38
uvicorn==0.34.0
